backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
requests
pycryptodome
aiofiles